+++ NEWS.md
@@ -4,6 +4,7 @@
 ### version 0.7.1, 2019/08/11
   * improvements in crossv(): confusion matrix
   * new funcion performance.measures(), providing recall, precision etc.
+  * new function text.size.penalize() to assess minimal sample size
   
 
 ### version 0.7.0, 2019/01/22